+++ crates/spk-cli/cmd-convert/src/cmd_convert.rs
@@ -18,8 +18,8 @@ pub struct Convert {
     #[clap(flatten)]
     pub requests: flags::Requests,
 
-    #[clap(short, long, global = true, parse(from_occurrences))]
-    pub verbose: u32,
+    #[clap(short, long, global = true, action = clap::ArgAction::Count)]
+    pub verbose: u8,
 
     #[clap(flatten)]
     pub formatter_settings: flags::DecisionFormatterSettings,
