+++ crates/spk-cli/common/src/error.rs
@@ -76,7 +76,7 @@ impl From<&str> for Error {
 }
 
 impl FormatError for Error {
-    fn format_error(&self, verbosity: u32) -> String {
+    fn format_error(&self, verbosity: u8) -> String {
         let mut msg = String::new();
         match self {
             /*
